+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "identifiable"
+require "encoded_ids"
 
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|